Moved HTTP-with-timeout code to a standalone helper class.
