+++ model.py
@@ -108,6 +108,7 @@ from util import (
     MetadataSimilarity,
     TitleProcessor,
 )
+from util.http import HTTP
 from util.permanent_work_id import WorkIDCalculator
 from util.summary import SummaryEvaluator
 
@@ -134,28 +135,7 @@ def production_session():
     return SessionManager.session(url)
 
 class PolicyException(Exception):
-    pass
-
-class RequestTimedOut(Exception):
-
-    def __init__(self, url, response):
-        self.url = url
-        self.hostname = urlparse.urlparse(url).netloc
-
-    def as_problem_detail_document(self, debug):
-        template = "Connection timed out while accessing %s"
-        if debug:
-            message = template % self.url
-        else:
-            message = template % self.hostname
-        if debug:
-            instance = self.url
-        else:
-            instance = None
-        return INTEGRATION_ERROR.detail(
-            detail=message, title="Timeout", instance=instance
-        )
-    
+    pass    
 
 class BaseMaterializedWork(object):
     """A mixin class for materialized views that incorporate Work and Edition."""
@@ -6300,37 +6280,17 @@ class Representation(Base):
         return json.dumps(dict(d))
 
     @classmethod
-    def request_with_timeout(cls, m, url, *args, **kwargs):
-        """Call the given request method and turn a timeout into a
-        RequestTimedOut exception.
-        """
-        if not 'timeout' in kwargs:
-            kwargs['timeout'] = 20
-        kwargs['timeout'] = 0.001
-        try:
-            response = m(url, *args, **kwargs)
-        except requests.exceptions.Timeout, e:
-            # Wrap the requests-specific Timeout exception 
-            # in a generic RequestTimedOut exception.
-            raise RequestTimedOut(url, e.message)
-        return response
-
-    @classmethod
     def simple_http_get(cls, url, headers, **kwargs):
         """The most simple HTTP-based GET."""
         if not 'allow_redirects' in kwargs:
             kwargs['allow_redirects'] = True
-        response = cls.request_with_timeout(
-            requests.get, url, headers=headers, **kwargs
-        )
+        response = HTTP.get_with_timeout(url, headers=headers, **kwargs)
         return response.status_code, response.headers, response.content
 
     @classmethod
     def simple_http_post(cls, url, headers, **kwargs):
         """The most simple HTTP-based POST."""
-        response = cls.request_with_timeout(
-            requests.post, url, headers=headers, **kwargs
-        )
+        response = HTTP.post_with_timeout(url, headers=headers, **kwargs)
         return response.status_code, response.headers, response.content
 
     @classmethod
